+++ rclc/include/rclc/executor_handle.h
@@ -22,6 +22,7 @@ extern "C"
 #endif
 
 #include <rcl/rcl.h>
+#include <rclc/visibility_control.h>
 
 /// TODO (jst3si) Where is this defined? - in my build environment this variable is not set.
 // #define ROS_PACKAGE_NAME "rclc"
@@ -56,6 +57,15 @@ typedef enum
 
 /// Type definition for callback function.
 typedef void (* rclc_callback_t)(const void *);
+
+/// Type definition for subscription callback function
+/// - incoming message
+// typedef void (* rclc_subscription_callback_t)(const void *);
+
+/// Type definition for subscription callback function
+/// - incoming message
+/// - additional callback context
+typedef void (* rclc_subscription_callback_with_context_t)(const void *, void *);
 
 /// Type definition for client callback function
 /// - request message
@@ -113,8 +123,8 @@ typedef struct
   /// only for service - ptr to response message
   void * data_response_msg;
 
-  /// only for service - ptr to additional service context
-  void * service_context;
+  /// ptr to additional callback context
+  void * callback_context;
 
   // TODO(jst3si) new type to be stored as data for
   //              service/client objects
@@ -128,6 +138,7 @@ typedef struct
   /// Storage for callbacks
   union {
     rclc_callback_t callback;
+    rclc_subscription_callback_with_context_t subscription_callback_with_context;
     rclc_service_callback_t service_callback;
     rclc_service_callback_with_request_id_t service_callback_with_reqid;
     rclc_service_callback_with_context_t service_callback_with_context;
@@ -183,6 +194,7 @@ typedef struct
  * \param[inout] handle_counters preallocated rclc_executor_handle_counters_t
  * \return `RCL_RET_INVALID_ARGUMENT` if `handle_counters` is a null pointer
  */
+RCLC_PUBLIC
 rcl_ret_t
 rclc_executor_handle_counters_zero_init(rclc_executor_handle_counters_t * handle_counters);
 
@@ -206,6 +218,7 @@ rclc_executor_handle_counters_zero_init(rclc_executor_handle_counters_t * handle
  * \return `RCL_RET_OK` if the \p handle was initialized successfully
  * \return `RCL_RET_INVALID_ARGUMENT` if \p h is a null pointer
  */
+RCLC_PUBLIC
 rcl_ret_t
 rclc_executor_handle_init(
   rclc_executor_handle_t * handle,
@@ -229,6 +242,7 @@ rclc_executor_handle_init(
  * \return `RCL_RET_OK` if \p h was cleared successfully
  * \return `RCL_RET_INVALID_ARGUMENT` if \p h is a null pointer
  */
+RCLC_PUBLIC
 rcl_ret_t
 rclc_executor_handle_clear(
   rclc_executor_handle_t * handle,
@@ -249,6 +263,7 @@ rclc_executor_handle_clear(
  * \return `RCL_RET_OK` if the handle was printed successfully
  * \return `RCL_RET_INVALID_ARGUMENT` if \p h is a null pointer
  */
+RCLC_PUBLIC
 rcl_ret_t
 rclc_executor_handle_print(rclc_executor_handle_t * handle);
 
@@ -270,6 +285,7 @@ rclc_executor_handle_print(rclc_executor_handle_t * handle);
  * \return pointer to the rcl-handle (rcl_subscription_t or rcl_timer_t)
  * \return NULL, if handle is a NULL pointer.
  */
+RCLC_PUBLIC
 void *
 rclc_executor_handle_get_ptr(rclc_executor_handle_t * handle);