+++ rclc_examples/src/example_client_node.c
@@ -1,5 +1,5 @@
 // Copyright (c) 2020 - for information on the respective copyright owner
-// see the NOTICE file and/or the repository https://github.com/micro-ROS/rclc.
+// see the NOTICE file and/or the repository https://github.com/ros2/rclc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,7 +19,6 @@
 #include "example_interfaces/srv/add_two_ints.h"
 
 #include <stdio.h>
-#include <unistd.h>
 
 #define RCCHECK(fn) {rcl_ret_t temp_rc = fn; if ((temp_rc != RCL_RET_OK)) {printf( \
         "Failed status on line %d: %d. Aborting.\n", __LINE__, (int)temp_rc); return 1;}}
@@ -71,10 +70,13 @@ int main(int argc, const char * const * argv)
   req.a = 24;
   req.b = 42;
 
-  sleep(2);   // Sleep a while to ensure DDS matching before sending request
+  rclc_sleep_ms(2000);   // Sleep a while to ensure DDS matching before sending request
 
   RCCHECK(rcl_send_request(&client, &req, &seq))
   printf("Send service request %ld + %ld.\n", req.a, req.b);
+
+  // Optional prepare for avoiding allocations during spin
+  rclc_executor_prepare(&executor);
 
   rclc_executor_spin(&executor);