+++ rclc_examples/src/example_lifecycle_node.c
@@ -15,7 +15,6 @@
 // limitations under the License.
 
 #include <stdio.h>
-#include <unistd.h>
 
 #include <rcl/error_handling.h>
 #include <rcutils/logging_macros.h>
@@ -89,7 +88,8 @@ int main(int argc, const char * argv[])
     &lifecycle_node,
     &my_node,
     &state_machine_,
-    &allocator);
+    &allocator,
+    true);
   if (rc != RCL_RET_OK) {
     printf("Error in creating lifecycle node.\n");
     return -1;