+++ rclc_lifecycle/CHANGELOG.rst
@@ -1,6 +1,26 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package rclc_lifecycle
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+3.0.1 (2021-07-17)
+------------------
+* Added quality declaration statement
+
+2.0.0 (2021-04-23)
+------------------
+* added codecov support
+* new API of rcl_lifecycle in Rolling required major version bump
+
+1.0.1 (2021-03-29)
+------------------
+* Windows port
+* Compatibility sleep function (Windows, POSIX-OS)
+* Fixed RCL lifecycle API change for Rolling
+
+1.0.0 (2021-03-04)
+------------------
+* Updated version
+
 0.1.7 (2021-01-20)
 ------------------
 * Updated version